fixed image_speed issue?

+++ trunk/ENIGMAsystem/SHELL/Universal_System/graphics_object.h
@@ -53,7 +53,7 @@ namespace enigma
   {
     //Sprites: these are mostly for higher tiers...
       int sprite_index;
-      int image_index;
+      double image_index;
       double image_speed;
       
       int image_number; //read only
